topology: imx8: Change DAI name
For now we are using generic simple-card as a machine driver, and
the DAI name must followa certain pattern <dai-name>-<codec-name>.

Temporarily change the DAI name in topology to match the name expected
by the machine driver.

There are still more changes to be done on the Linux kernel until
the name gets stable, so do this temporary change now.

Will push the code upstream once the Linux kernel changes are stable.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>